requirements.txt
django
pyenchant